move wechaty & wechaty-puppet to /src
