+++ src/wechaty-puppet/file_box.go
