+++ src/wechaty-puppet/schemas/image.go
